[Feat/gallery]: 뷰 모델에 페이징 데이터 불러오는 기능 구현
